Fixed a bug that occurs if the first line is continued.
The bug was that the while loop looking for the first continued line in a series of continued lines, did not take into account that you can't go (forward-line -1) at the beginning of the buffer.

The bug is fixed by adding a check for (not (bobp)) in the while condition.